Added a dm check

+++ commands/delete.js
@@ -8,24 +8,26 @@ module.exports = {
   description: 'löscht Nachrichten',
   execute(message, args) {
 
+    if (message.guild === null) return;
+
     var args = message.content.slice(6).split(" ");
 
     const löschAnzahl = args[1];
 
-    if(!löschAnzahl || löschAnzahl == 0 || löschAnzahl > 100){
+    if (!löschAnzahl || löschAnzahl == 0 || löschAnzahl > 100) {
 
       message.reply("Please provide a number between 2 and 100!");
 
-    } else if (message.author.id == message.guild.ownerID){
+    } else if (message.author.id == message.guild.ownerID) {
 
       message.delete;
 
       return message.channel.bulkDelete(löschAnzahl);
 
-    } else{
+    } else {
       message.reply("Youre not in the owner!");
 
-      console.log("User " + message.author.id+ " hat versucht den command `delete` zu benutzen");
+      console.log("User " + message.author.id + " hat versucht den command `delete` zu benutzen");
     }
   }
 }
